% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwm.R
\name{kwm}
\alias{kwm}
\alias{predict.kwm}
\title{Produce a model object based on regular expressions.}
\usage{
kwm(include = character(), exclude = character(), varname,
  grepl_opts = NULL)

\method{predict}{kwm}(object, newdata, progress = interactive(),
  return_names = FALSE, ...)
}
\arguments{
\item{include}{Character. List of regular expression patterns, any one of which will return a positive match.}

\item{exclude}{Character. List of regular expression patterns, any one of which will return a negative match, overriding any other matches with \code{include}}

\item{varname}{Character. A character vector of length one with the column name to be tested.}

\item{grepl_opts}{Named list of arguments to pass to \link[base:grepl]{base::grepl}.}

\item{object}{A \code{kwm} model object}

\item{newdata}{A data.frame containing a text column}

\item{progress}{Logical. Display a progress bar?}

\item{return_names}{Logical. Add original text as names to the resulting
logical vector?}

\item{...}{Other arguments passed from functions. Not used currently.}
}
\value{
\code{kwm} returns a model object of class \code{kwm}

\code{predict.kwm} reutrns a logical vector.
}
\description{
Produce a model object based on regular expressions.
}
