% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwm.R
\name{kwm}
\alias{kwm}
\title{Produce a model object based on regular expressions}
\usage{
kwm(include_list, exclude_list = character(), varname)
}
\arguments{
\item{varname}{Character. A character vector of length one with the column name to be tested.}

\item{include}{Character. List of regular expression patterns, any one of which will return a positive match.}

\item{exclude}{Character. List of regular expression patterns, any one of which will return a negative match, overriding any other matches with \code{include}}
}
\value{
A moel object of class \code{kwm_model}
}
\description{
Produce a model object based on regular expressions
}
